RM#30605: Fix message service use long instead of integer as parameter for id

+++ axelor-base/src/main/java/com/axelor/apps/base/service/message/MessageServiceBaseImpl.java
@@ -68,7 +68,7 @@ public class MessageServiceBaseImpl extends MessageServiceImpl {
   @Transactional
   public Message createMessage(
       String model,
-      int id,
+      long id,
       String subject,
       String content,
       EmailAddress fromEmailAddress,
